+++ Com.GleekFramework.ContractSdk/Params/PageParam.cs
@@ -15,7 +15,7 @@ namespace Com.GleekFramework.ContractSdk
         /// </summary>
         [FromQuery(Name = "next_id")]
         [JsonProperty("next_id"), JsonPropertyName("next_id")]
-        public long? NextId { get; set; }
+        public decimal? NextId { get; set; }
 
         /// <summary>
         /// 分页大小
